Aesthetics finished
Recoloured everything. Also changed how the calculator resizes quite substantialy.

Felt that the smaller version was too similar to the larger one. Made it smaller and fixed proportions.

There is currently a bug where a long string of numbers will spill out of the calculator's screen. That is the last thing to take care of here.